[Refactor][1/N] Cleanup torchax path and consolidate vllm files
